pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "beam_abm"
version = "0.1.0"
description = "LLM-based Agent-Based Model for behavioral analysis"
authors = [{ name = "Adam Amer", email = "adam@example.com" }]
readme = "README.md"
requires-python = ">=3.13"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    # Core
    "cruft>=2.16.0",
    "datamodel-code-generator>=0.35.0",
    "ipywidgets>=8.1.8",
    "jupyterlab>=4.4.10",
    "python-dotenv>=1.2.1",
    "python-levenshtein>=0.27.3",
    "scikit-learn>=1.7.2",
    "shap>=0.49.1",
    "tensorboard>=2.20.0",
    "torch>=2.8.0",
    "torchinfo>=1.8.0",
    # Dataset processing
    "networkx>=3.5",
    "pandas>=2.3.3",
    "xgboost>=3.1.1",
    #"ydata-profiling>=4.17.0",
    # DataFrames
    "polars>=1.35.1",
    # CLI
    "typer>=0.20.0",
    # LLM Behaviour
    "tiktoken>=0.12.0",
    "xformers>=0.0.32.post1",
    "editables>=0.5",
    ## Google/OpenAI
    "google-cloud-aiplatform>=1.124.0",
    "instructor>=1.12.0",
    "openai>=1.0.0",
    # Logging
    "loguru>=0.7.3",
    "huggingface-hub[hf-xet]>=0.36.0",
    "statsmodels>=0.14.2",
    "pyarrow>=22.0.0",
    "nbformat>=5.10.4",
    "notebook>=7.4.7",
    "msgspec>=0.19.0",
    "optuna>=4.5.0",
    "vllm>=0.13.0",
    "matplotlib>=3.10.7",
    "seaborn>=0.13.2",
    # ABM
    "numpy>=1.26.0",
    "scipy>=1.14.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "mesa-frames",
    "jupyter-cache>=1.0.1",
]

[dependency-groups]
dev = [
    "ruff>=0.14.3",
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "mkdocs-material>=9.6.23",
    "mkdocstrings[python]>=0.30.1",
    "pre-commit>=4.3.0",
    "beartype>=0.22.5",
    "python-dotenv>=1.2.1",
    "cruft>=2.16.0",
    "hatchling>=1.27.0",
]

paper = [
    "marimo>=0.8.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "plotly>=5.0.0",
]

cuda = ["flashinfer-python>=0.2.7.post1", "flash-attn>=2.8.1"]


[project.urls]
Homepage = "https://github.com/adamamer20/beam-abm"
Documentation = "https://adamamer20.github.io/beam-abm/"
Repository = "https://github.com/adamamer20/beam-abm"
Issues = "https://github.com/adamamer20/beam-abm/issues"

[tool.uv]
no-build-isolation-package = ["flash-attn"]
default-groups = ["dev"]

[tool.uv.sources]
torch = [
    { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
    { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchaudio = [
    { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
xformers = [
    { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
mesa-frames = { git = "https://github.com/projectmesa/mesa-frames", branch = "main" }

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "RUF100", # unused noqa
]
ignore = [
    "E501", # line too long, handled by black
    "B008", # do not perform function calls in argument defaults
    "C901", # too complex
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"tests/**/*" = ["B018"]

[tool.ruff.lint.isort]
known-first-party = ["beam_abm"]

[tool.pytest.ini_options]
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "unit: marks tests as unit tests",
    "integration: marks tests as integration tests",
    "gpu: marks tests as requiring GPU and real models (local development only)",
]

[tool.coverage.run]
source = ["src"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.hatch.version]
path = "src/beam_abm/__init__.py"

[tool.hatch.build.targets.wheel]
packages = ["src/beam_abm"]

[tool.codespell]
ignore-words-list = "codespell-ignore-words.txt"
